RPM build caching
A new `-c` option will enable rpm build caching per spec. This is
currently only suitable for vendor packages, because it determines
changes by the hash of the spec file only

This mechanism can be extended later to specs that depend on source
packages by hashing their contents.

To use this, you need to add the following to your RPM build
Dockerfiles:

    @IF [ ! -z "$BUILDER_CACHE" ]
    @EVAL ADD builder/cache/${BUILDER_TARGET}/ /cache/old/
    ENV BUILDER_CACHE 1
    @ENDIF

and set `BUILDER_CACHE_THIS=1` on the build-specs.sh lines where
you want to enable this.